refactor(load): evict the warehouse-load orchestrator from cli/dispatch.rs
Arch-roast Strong (challenge-confirmed): dispatch.rs's own doc promises "every
arm is a thin adapter... call exactly one function" and 12 of 13 commands
honored it — Load embedded the whole ~800-line orchestrator in the router
(plan resolution, active-run/GC guards, the per-mode load drivers
full/incremental/CDC, the LoadCtx ledger, failure aggregation).

Moved WHOLESALE to src/load/orchestrate.rs beside the layer it drives; the
Load arm is one call again (load::orchestrate::run_loads). dispatch.rs:
1952 → 899 lines. Entry points (run_loads, LoadArgs, resolve_rivet_bin) are
pub — the load module compiles into both targets and the caller lives in the
bin, so the lib target needs the public root to see the code as live.

Tests moved with their subjects (load_ledger_tests, rivet_bin_tests); the
empty_loads_message cases stayed beside their fn in dispatch (it belongs to
`rivet state loads`, not the load orchestrator).

One cop retargeted, made stronger not weaker: destructive_delete_gate now
reads BOTH orchestrate.rs (where its subjects — cleanup_target,
maybe_gc_orphans — live) AND dispatch.rs, so a delete consumer left in or
returned to the router is still graded.

Verbatim move otherwise. Green: 2517 lib · 265 offline · 12 live load.
